Add new scripts and functions
