+++ .github/scripts/src/getDate.ts
@@ -1,0 +1,3 @@
+export function getDate(now: Date): string {
+  return `${now.getFullYear()}${now.getMonth() + 1}${now.getDate()}`;
+}
