+++ .github/scripts/src/getTime.ts
@@ -1,0 +1,3 @@
+export function getTime(now: Date): string {
+  return `${now.getHours()}${now.getMinutes()}${now.getSeconds()}`;
+}
