+++ .github/scripts/src/npmBuild.ts
@@ -1,0 +1,6 @@
+import { workspace } from "src";
+import { execWrapper } from "./execWrapper";
+
+export async function npmBuild(cwd = workspace) {
+  await execWrapper("npm", ["run", "build"], { cwd });
+}
